feat(server): allow custom message stringification and parsing
If you want to use a different transport than JSON over the websocket, or use an extension of JSON to support things like bigints, it's handy to be able to control the stringification and unstringifiation of the messages going across this socket! This adds options to the server to allow overriding the `stringifyMessage` and `parseMessage` functions the server uses so userland can hook in.