feat: allow custom JSON message stringification and parsing
If you want to use a fancier types inside your JSON over the websocket like bigints or dates, it's handy to be able to control the stringification and unstringifiation of the messages going across this socket! This adds options to the server to allow overriding the `revive` and `replace` functions the server uses in JSON.stringify and JSON.parse so userland can hook in and alter the behaviour. We stay compatible with the graphql-ws protocol by sticking with JSON but still allow flexibility to marshal fancier types to and from JSON easily.