Fix behaviour around closing the socket if the onConnect handler throws
The JSDoc comment for the onConnect handler outlines behaviour that makes sense to me -- if the context can't be set up, we close the socket and send an error to the client. I couldn't find anything that actually implemented that though, so I added the behaviour. My test was failing before I added this change, so I think maybe the behaviour was just lost at some point.