feat(tests): improve test client wait methods
The `waitForMessage` function will now throw an error if the WebSocket is closed
before a message comes in. Previously, the call would continue to wait until the
test timeout is was exceeded.

Similarly, the `waitForClose` function will now throw an error if the WebSocket
receives a message before closing. Previously, any messages received while waiting
for the WebSocket to close would be ignored, regardless of whether the messages
were expected or not.