feat(tests): expand error handling tests
This commit adds new tests to cover how errors from the `resolve` and `subscribe`
functions on `GraphQLFieldConfig` are handled, for both queries and subscriptions.
(Since mutations follow the same code path as queries, tests for them haven’t been
included.)

For queries, the test matrix is pretty simple; the `resolve` function either returns
a result or throws an error. The `subscribe` function isn’t called.

For subscriptions, the test matrix gets more complicated. The `subscribe` method can
either throw an error or return a generator. The generator can in turn either return
a result or throw an error from its `next` method, on any invocation. In addition,
the `resolve` method can either return a result or throw an error.

To accommodate testing all these combinations, a new query and a new subscription
are added to the fixture schema. Both are called `throwingFrom`, and take parameters
to allow the caller to control from where errors are thrown.

Tests for the full matrix have been added to both the client and server side tests,
where the event stream from each case is serialized to an object array and then
compared against a snapshot.